viedeos_manager-New_code_with_resolve_issoe/requirements_get.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-countries==6.1.2
django-filter==2.2.0
django-import-export==2.1.0
django-paypal==1.0.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2018.7
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
